shortener_app/requirements.txt
Pillow
PyJWT
Requests
SQLAlchemy
aiofiles
aiohttp
alembic
beautifulsoup4
fastapi
httpx
itsdangerous
jinja2
playwright
psycopg2-binary
pydantic_settings
pyjwt 
pyotp
pydantic_settings
python-dotenv
qrcodegen
requests
slowapi
starlette
uvicorn
validators